+++ src/components/functional/pending-transactions-panel/pending-transactions-panel.types.ts
@@ -5,10 +5,3 @@ export enum PendingTransactionsEventsEnum {
   OPEN_PENDING_TRANSACTIONS_PANEL = 'OPEN_PENDING_TRANSACTIONS_PANEL',
   DATA_UPDATE = 'DATA_UPDATE',
 }
-
-export interface IPendingTransactionsPanelData {
-  isPending: boolean;
-  title: string;
-  subtitle?: string;
-  shouldClose?: boolean;
-}